restaurant_picker/requirements.txt
asgiref==3.7.2
attrs==23.1.0
awsebcli==3.20.9
bcrypt==4.0.1
blessed==1.20.0
botocore==1.31.44
cement==2.8.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.3
cryptography==41.0.3
distro==1.8.0
Django==4.2.5
django-extensions==3.2.3
docker==6.1.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.4
jmespath==1.0.1
jsonschema==3.2.0
packaging==23.1
paramiko==3.3.1
pathspec==0.10.1
psycopg2==2.9.7
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
PyYAML==5.4.1
requests==2.31.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
texttable==1.6.7
typing_extensions==4.7.1
urllib3==1.26.16
wcwidth==0.1.9
websocket-client==0.59.0
